pyproject.toml
[tool.poetry]
name = "httpy"
version = "0.0.1"
description = "HTTP client with native yaml support."
authors = ["Gabriel <gabriel59kg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
httpx = "^0.17"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
pre-commit = "^2.10.1"
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-icdiff = "^0.5"
respx = "^0.16.3"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
